move task functions to hospital model
