+++ rp_asos/models.py
@@ -1,5 +1,7 @@
 from django.db import models
 
+from sidekick import utils
+
 from rp_redcap.models import Project
 
 
@@ -20,6 +22,54 @@ class Hospital(models.Model):
 
     def __str__(self):
         return "{} - {}".format(self.project.name, self.name)
+
+    def create_hospital_wa_group(self):
+        if not self.whatsapp_group_id:
+            # WA group subject is limited to 25 characters
+            self.whatsapp_group_id = utils.create_whatsapp_group(
+                self.project.org, "{} - ASOS2".format(self.name[:17])
+            )
+            self.save()
+
+    def get_wa_group_info(self):
+        group_info = utils.get_whatsapp_group_info(
+            self.project.org, self.whatsapp_group_id
+        )
+        group_info["id"] = self.whatsapp_group_id
+        return group_info
+
+    def send_group_invites(self, group_info, wa_ids):
+        invites = []
+        for wa_id in wa_ids:
+            if wa_id not in group_info["participants"]:
+                invites.append(wa_id)
+
+        if invites:
+            invite_link = utils.get_whatsapp_group_invite_link(
+                self.project.org, group_info["id"]
+            )
+            for wa_id in invites:
+                utils.send_whatsapp_template_message(
+                    self.project.org,
+                    wa_id,
+                    "whatsapp:hsm:npo:praekeltpbc",
+                    "asos2_notification2",
+                    {
+                        "default": "Hi, please join the ASOS2 Whatsapp group: {}".format(
+                            invite_link
+                        )
+                    },
+                )
+
+    def add_group_admins(self, group_info, wa_ids):
+        for wa_id in wa_ids:
+            if (
+                wa_id in group_info["participants"]
+                and wa_id not in group_info["admins"]
+            ):
+                utils.add_whatsapp_group_admin(
+                    self.project.org, group_info["id"], wa_id
+                )
 
 
 class PatientRecord(models.Model):
